test: dataset upload component tests
