+++ src/app/modules/dataset-upload-module/__test__/dataset/add-dataset.test.tsx
@@ -1,20 +1,24 @@
-import {
-  fireEvent,
-  render,
-  screen,
-  waitFor,
-  waitForElementToBeRemoved,
-} from "@testing-library/react";
-import AddDatasetFragment, {
-  DropZone,
-} from "app/modules/dataset-upload-module/upload-steps/addDatasetFragment";
+import { fireEvent, render, screen, waitFor } from "@testing-library/react";
+import AddDatasetFragment from "app/modules/dataset-upload-module/upload-steps/addDatasetFragment";
 import userEvent from "@testing-library/user-event";
 import { StoreProvider, createStore } from "easy-peasy";
 import {
   AuthTokenModel,
   AuthTokenState,
 } from "app/state/api/action-reducers/sync";
-import { mockFileRejections } from "../mock-data";
+import axios from "axios";
+
+const mockOpenPicker = jest.fn();
+jest.mock("react-google-drive-picker", () => {
+  return {
+    __esModule: true,
+    default: () => {
+      return [mockOpenPicker];
+    },
+  };
+});
+
+jest.mock("axios");
 
 const appFn = (
   mockSetSelectedFile: jest.Mock<any, any, any>,
@@ -27,7 +31,7 @@ const appFn = (
   return (
     <StoreProvider store={mockStore}>
       <AddDatasetFragment
-        setFile={mockSetSelectedFile}
+        onFileSubmit={mockSetSelectedFile}
         disabled={false}
         processingError={false}
         setIsExternalSearch={mockSetIsExternalSearch}
@@ -44,10 +48,11 @@ test("local upload of dataset", async () => {
 
   const file = new File(["(⌐□_□)"], "chucknorris.csv", { type: "text/csv" });
   const dndText = screen.getByText(/drag and drop/i);
+  expect(dndText).toBeInTheDocument();
   const uploadButton = screen.getByTestId("local-upload") as HTMLInputElement;
 
-  expect(dndText).toBeInTheDocument();
   Object.defineProperty(uploadButton, "files", { value: [file] });
+
   fireEvent.drop(uploadButton);
 
   await waitFor(() => {
@@ -71,46 +76,33 @@ test("external search of dataset", async () => {
 });
 
 test("google drive button", async () => {
-  const user = userEvent.setup();
-  const acceptedFiles = [
-    new File(["(⌐□_□)"], "chucknorris.csv", { type: "text/csv" }),
-  ];
-
-  const mockGetRootProps = jest.fn();
-  const mockGetInputProps = jest.fn();
+  const mockSetSelectedFile = jest.fn();
   const mockSetIsExternalSearch = jest.fn();
-  const mockHandleOpenPicker = jest.fn();
+  const app = appFn(mockSetSelectedFile, mockSetIsExternalSearch);
+  mockOpenPicker.mockImplementation(({ callbackFunction }) => {
+    callbackFunction({ docs: [{ id: "123", name: "test.csv", type: "file" }] });
+  });
+  (axios.get as jest.Mock).mockResolvedValueOnce({ data: "access_token" });
 
-  render(
-    <DropZone
-      disabled={false}
-      getRootProps={mockGetRootProps}
-      getInputProps={mockGetInputProps}
-      isDragActive={false}
-      fileRejections={mockFileRejections}
-      acceptedFiles={acceptedFiles}
-      handleOpenPicker={mockHandleOpenPicker}
-      uploadError={false}
-      setIsExternalSearch={mockSetIsExternalSearch}
-    />
-  );
-
+  render(app);
   const googleDriveButton = screen.getByTestId("google-drive-button");
-  await user.click(googleDriveButton);
-  expect(mockHandleOpenPicker).toHaveBeenCalled();
+
+  expect(googleDriveButton).toBeInTheDocument();
+  await userEvent.click(googleDriveButton);
+  (axios.get as jest.Mock).mockResolvedValueOnce({ data: [] });
+
+  // Assert that openPicker is called with the correct parameters
+  expect(mockOpenPicker).toHaveBeenCalledWith({
+    clientId: expect.any(String),
+    developerKey: expect.any(String),
+    viewId: "SPREADSHEETS",
+    supportDrives: true,
+    token: "access_token",
+    setSelectFolderEnabled: true,
+    callbackFunction: expect.any(Function),
+  });
+
+  // await waitFor(() => {
+  //   expect(mockSetSelectedFile).toHaveBeenCalled();
+  // });
 });
-
-// test("google drive picker", async () => {
-//   const mockSetSelectedFile = jest.fn();
-//   const mockSetIsExternalSearch = jest.fn();
-
-//   const app = appFn(mockSetSelectedFile, mockSetIsExternalSearch);
-
-//   render(app);
-
-//   const googleDriveButton = screen.getByTestId("google-drive-button");
-//   fireEvent.click(googleDriveButton);
-
-//   await waitForElementToBeRemoved(() => screen.getByText(/drag and drop/i));
-//   expect(screen.getByText(/select a file/i)).toBeInTheDocument();
-// });
